Fix include compilation issue
